chore(type-designer): move the plugin from uilib to a dedicated package inside plugins
