+++ packages/plugins/type-designer/README.md
@@ -1,3 +1,3 @@
-# Type Designer
+# Type Designer Plugin
 
-Refer to [TypeDesigner UiLib folder](../../uilib/src/lib/plugins/type-designer/) to see the source code.
+The Type Designer plugin for OpenSCD simplifies the creation and management of types elements in SCL file. It features a board with six parts (Substation, Voltage Level, Bay, IED, Logical Device, Logical Nodes), drag-and-drop functionality for creating type references, and a parameter drawer for editing element details.
