+++ packages/plugins/type-designer/index.html
@@ -1,0 +1,12 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>OpenSCD Plugin - TypeDesigner</title>
+  </head>
+  <body>
+    <script type="module" src="/src/plugin.ts"></script>
+  </body>
+</html>
