+++ packages/plugins/type-designer/src/components/elements-type-container/element-type-container.svelte
