+++ packages/plugins/type-designer/src/stores/data-types-templates.store.ts
