+++ packages/plugins/type-designer/src/stores/xml-document.store.ts
