+++ packages/uilib/package.json
@@ -38,7 +38,6 @@
 		"svelte-material-ui": "7.0.0-beta.8"
 	},
 	"devDependencies": {
-		"@oscd-plugins/types": "workspace:^",
 		"@playwright/test": "^1.28.1",
 		"@sveltejs/adapter-auto": "^1.0.0",
 		"@sveltejs/kit": "^1.5.0",
