+++ packages/uilib/src/lib/plugins/type-designer/+page.svelte
@@ -1,12 +1,12 @@
-<Example name="Type Designer">
+<!-- <Example name="Type Designer">
 	<TypeDesignerPlugin xmlDocument={doc} />
 </Example>
 
 <script lang="ts">
-import TypeDesignerPlugin from './type-designer-plugin.svelte'
+import TypeDesignerPlugin from '@oscd-plugins/type-designer'
 import { xmlStr } from '../../../test-files/simple_v5'
 import { Example } from '../../components/internal'
 
 const parser = new DOMParser()
 const doc = parser.parseFromString(xmlStr, 'text/xml') as unknown as XMLDocument
-</script>
+</script> -->
