--- packages/uilib/src/lib/plugins/type-designer/index.ts
@@ -1,1 +1,0 @@
-export { default as TypeDesignerPlugin } from './type-designer-plugin.svelte'
